Issue 78: Harmonize fixtures Fix a bug on folder repository
