+++ docker-compose.override.yaml.dist
@@ -11,4 +11,4 @@ services:
 
     mysql:
         ports:
-            - '33006:33006'
+            - '33006:3306'
