+++ features/documents/display.feature
@@ -11,15 +11,15 @@ Feature: Display files and folders information
     And I press "Connexion"
     And I am on "documents"
     And I follow "Ajouter un fichier"
-    And I attach the file "features/Context/fixtures/black_cat.jpg" to "file_filePath"
+    And I attach the file "tests/fixtures/images/black_cat.jpg" to "file_filePath"
     And I press "file_submit"
     And I follow "A folder at root"
     And I follow "Ajouter un fichier"
-    And I attach the file "features/Context/fixtures/white_cat.jpg" to "file_filePath"
+    And I attach the file "tests/fixtures/images/white_cat.jpg" to "file_filePath"
     And I press "file_submit"
     And I follow "Another folder inside a folder"
     And I follow "Ajouter un fichier"
-    And I attach the file "features/Context/fixtures/siamese_cat.jpg" to "file_filePath"
+    And I attach the file "tests/fixtures/images/siamese_cat.jpg" to "file_filePath"
     And I press "file_submit"
 
   Scenario: I can see files and folders information
