+++ tests/features/user/administrators/edit.feature
@@ -1,4 +1,4 @@
-@end-to-end
+@end-to-end @fixtures-users
 Feature: Edit user profiles
   In order to administrate user accounts
   As an administrator
