+++ tests/features/user/administrators/list.feature
@@ -1,4 +1,4 @@
-@end-to-end
+@end-to-end @fixtures-users
 Feature: Manage user accounts
   In order to administrate user accounts
   As an administrator
