+++ tests/features/user/administrators/roles.feature
@@ -1,4 +1,4 @@
-@end-to-end
+@end-to-end @fixtures-users
 Feature: Change users roles
   In order to administrate user accounts
   As an administrator
