+++ tests/features/user/anonymous/resetting.feature
@@ -1,4 +1,4 @@
-@end-to-end
+@end-to-end @fixtures-users
 Feature: Réinitialiser le mot de passe
   In order to access the application
   As an anonymous user
