+++ tests/fixtures/images/black_cat.jpg
