+++ tests/fixtures/images/siamese_cat.jpg
