+++ tests/fixtures/images/white_cat.jpg
